+++ pkg/client/bievents.go
@@ -4,7 +4,9 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"math"
 	"net/http"
 	"sync"
 	"time"
@@ -12,18 +14,27 @@ import (
 	"github.com/gesundheitscloud/go-svc/pkg/logging"
 )
 
+type EventAttempts struct {
+	Data     []byte
+	Attempts int
+}
+
 type BIEventsFilter struct {
 	svcAddr   string
 	svcSecret string
 	caller    *caller
-	channel   chan []byte
+	channel   chan EventAttempts
 	wg        *sync.WaitGroup
 }
+
+const MaxAttempts = 5
+
+var ErrSendingFailed = errors.New("failed sending BI Event")
 
 var userAgentBIEventsFilter = "go-svc.client.BIEventsFilter"
 
 func NewBIEventsFilter(runCtx context.Context, svcAddr, svcSecret, caller string) (done chan struct{}, bic BIEventsFilter) {
-	c := make(chan []byte)
+	c := make(chan EventAttempts)
 	done = make(chan struct{})
 
 	bic = BIEventsFilter{
@@ -38,15 +49,21 @@ func NewBIEventsFilter(runCtx context.Context, svcAddr, svcSecret, caller string
 	// worker goroutine loop sending events to the bi service
 	go func() {
 		for {
-			data := <-c
-			err := bic.postEvent(data)
+			event := <-c
+			err := bic.postEvent(event)
+			event.Attempts++
 			if err != nil {
-				// let worker sleep to reduce cpu load
-				time.Sleep(time.Second)
-				// requeue data
-				go func(data []byte) {
-					bic.channel <- data
-				}(data)
+				// requeue event if attempted less than maximum times
+				if event.Attempts < MaxAttempts {
+					// let worker sleep with exponential backoff to reduce cpu load
+					time.Sleep(time.Second * time.Duration(math.Pow(2.0, float64(event.Attempts+1))))
+					go func(event EventAttempts) {
+						bic.channel <- event
+					}(event)
+				} else {
+					logging.LogErrorfCtx(context.Background(), ErrSendingFailed, "failed after %d attempts", MaxAttempts)
+					bic.wg.Done()
+				}
 			} else {
 				bic.wg.Done()
 			}
@@ -73,14 +90,14 @@ func (bic BIEventsFilter) Send(data interface{}) error {
 	if err != nil {
 		return err
 	}
-	go func(data []byte) {
+	go func(data EventAttempts) {
 		bic.channel <- data
-	}(json)
+	}(EventAttempts{json, 0})
 	return nil
 }
 
-func (bic BIEventsFilter) postEvent(data []byte) error {
+func (bic BIEventsFilter) postEvent(event EventAttempts) error {
 	contentURL := fmt.Sprintf("%s/api/v1/events", bic.svcAddr)
-	_, _, err := bic.caller.call(context.Background(), contentURL, "POST", bic.svcSecret, userAgentBIEventsFilter, bytes.NewBuffer(data), http.StatusOK)
+	_, _, err := bic.caller.call(context.Background(), contentURL, "POST", bic.svcSecret, userAgentBIEventsFilter, bytes.NewBuffer(event.Data), http.StatusOK)
 	return err
 }